add research's public date

+++ app/src/main/java/com/zakrodionov/protovary/data/entity/ResearchCompact.kt
@@ -31,5 +31,5 @@ data class ResearchCompact(
     var url: String?,
 
     @SerializedName("utime")
-    var utime: Int?
+    var utime: Long?
 ) : Serializable, DisplayableItem
